Remove the references to the string problems
This removes all the references to the string problems.

I also copied over a lot of the changes that I made to background parts of the string lab. This isn't great, and is a non-trivial kind of content duplication that may create problems down the road, but that's where we are for now.